requirements.txt
fastapi
uvicorn
pydantic
dnspython
pymongo
motor
python-dotenv
transformers
flask_cors

